+++ engine/api/src/main/java/io/deephaven/engine/table/MatchPair.java
@@ -4,7 +4,7 @@
 package io.deephaven.engine.table;
 
 import io.deephaven.api.ColumnName;
-import io.deephaven.api.agg.Pair;
+import io.deephaven.api.Pair;
 import io.deephaven.base.log.LogOutput;
 import io.deephaven.io.log.impl.LogOutputStringImpl;
 import io.deephaven.api.util.NameValidator;
