+++ table-api/src/main/java/io/deephaven/api/agg/AggregationOptimizer.java
@@ -4,6 +4,7 @@
 package io.deephaven.api.agg;
 
 import io.deephaven.api.ColumnName;
+import io.deephaven.api.Pair;
 import io.deephaven.api.agg.spec.AggSpec;
 
 import java.util.ArrayList;
