+++ table-api/src/main/java/io/deephaven/api/agg/AggregationOutputs.java
@@ -4,6 +4,7 @@
 package io.deephaven.api.agg;
 
 import io.deephaven.api.ColumnName;
+import io.deephaven.api.Pair;
 
 import java.util.Collection;
 import java.util.Objects;
