+++ table-api/src/main/java/io/deephaven/api/agg/spec/AggSpec.java
@@ -7,7 +7,7 @@ import io.deephaven.api.ColumnName;
 import io.deephaven.api.SortColumn;
 import io.deephaven.api.agg.Aggregation;
 import io.deephaven.api.agg.ColumnAggregation;
-import io.deephaven.api.agg.Pair;
+import io.deephaven.api.Pair;
 import io.deephaven.api.object.UnionObject;
 
 import java.util.Arrays;
