+++ table-api/src/main/java/io/deephaven/api/updateby/ColumnUpdateOperation.java
@@ -1,7 +1,7 @@
 package io.deephaven.api.updateby;
 
 import io.deephaven.annotations.BuildableStyle;
-import io.deephaven.api.agg.Pair;
+import io.deephaven.api.Pair;
 import io.deephaven.api.updateby.spec.UpdateBySpec;
 import org.immutables.value.Value.Immutable;
